Fix System.Environment implementation for uapaot
