+++ src/System.Runtime.Extensions/src/Resources/Strings.resx
@@ -240,87 +240,6 @@
   <data name="InvalidOperation_GetVersion" xml:space="preserve">
     <value>OSVersion's call to GetVersionEx failed</value>
   </data>
-  <data name="Argument_IllegalEnvVarName" xml:space="preserve">
-    <value>Environment variable name cannot contain equal character.</value>
-  </data>
-  <data name="Argument_LongEnvVarValue" xml:space="preserve">
-    <value>Environment variable name or value is too long.</value>
-  </data>
-  <data name="Argument_StringFirstCharIsZero" xml:space="preserve">
-    <value>The first char in the string is the null character.</value>
-  </data>
-  <data name="Argument_StringZeroLength" xml:space="preserve">
-    <value>String cannot be of zero length.</value>
-  </data>
-  <data name="Arg_RegSubKeyAbsent" xml:space="preserve">
-    <value>Cannot delete a subkey tree because the subkey does not exist.</value>
-  </data>
-  <data name="Arg_RegKeyDelHive" xml:space="preserve">
-    <value>Cannot delete a registry hive's subtree.</value>
-  </data>
-  <data name="Arg_RegKeyNoRemoteConnect" xml:space="preserve">
-    <value>No remote connection to '{0}' while trying to read the registry.</value>
-  </data>
-  <data name="Arg_RegKeyOutOfRange" xml:space="preserve">
-    <value>Registry HKEY was out of the legal range.</value>
-  </data>
-  <data name="Arg_RegKeyNotFound" xml:space="preserve">
-    <value>The specified registry key does not exist.</value>
-  </data>
-  <data name="Arg_RegKeyStrLenBug" xml:space="preserve">
-    <value>Registry key names should not be greater than 255 characters.</value>
-  </data>
-  <data name="Arg_RegValStrLenBug" xml:space="preserve">
-    <value>Registry value names should not be greater than 16,383 characters.</value>
-  </data>
-  <data name="Arg_RegBadKeyKind" xml:space="preserve">
-    <value>The specified RegistryValueKind is an invalid value.</value>
-  </data>
-  <data name="Arg_RegGetOverflowBug" xml:space="preserve">
-    <value>RegistryKey.GetValue does not allow a String that has a length greater than Int32.MaxValue.</value>
-  </data>
-  <data name="Arg_RegSetMismatchedKind" xml:space="preserve">
-    <value>The type of the value object did not match the specified RegistryValueKind or the object could not be properly converted.</value>
-  </data>
-  <data name="Arg_RegSetBadArrType" xml:space="preserve">
-    <value>RegistryKey.SetValue does not support arrays of type '{0}'. Only Byte[] and String[] are supported.</value>
-  </data>
-  <data name="Arg_RegSetStrArrNull" xml:space="preserve">
-    <value>RegistryKey.SetValue does not allow a String[] that contains a null String reference.</value>
-  </data>
-  <data name="Arg_RegInvalidKeyName" xml:space="preserve">
-    <value>Registry key name must start with a valid base key name.</value>
-  </data>
-  <data name="Arg_DllInitFailure" xml:space="preserve">
-    <value>One machine may not have remote administration enabled, or both machines may not be running the remote registry service.</value>
-  </data>
-  <data name="Arg_RegSubKeyValueAbsent" xml:space="preserve">
-    <value>No value exists with that name.</value>
-  </data>
-  <data name="Argument_InvalidRegistryOptionsCheck" xml:space="preserve">
-    <value>The specified RegistryOptions value is invalid.</value>
-  </data>
-  <data name="Argument_InvalidRegistryViewCheck" xml:space="preserve">
-    <value>The specified RegistryView value is invalid.</value>
-  </data>
-  <data name="Argument_InvalidRegistryKeyPermissionCheck" xml:space="preserve">
-    <value>The specified RegistryKeyPermissionCheck value is invalid.</value>
-  </data>
-  <data name="InvalidOperation_RegRemoveSubKey" xml:space="preserve">
-    <value>Registry key has subkeys and recursive removes are not supported by this method.</value>
-  </data>
-  <data name="ObjectDisposed_RegKeyClosed" xml:space="preserve">
-    <value>Cannot access a closed registry key.</value>
-  </data>
-  <data name="Security_RegistryPermission" xml:space="preserve">
-    <value>Requested registry access is not allowed.</value>
-  </data>
-  <data name="UnauthorizedAccess_RegistryKeyGeneric_Key" xml:space="preserve">
-    <value>Access to the registry key '{0}' is denied.</value>
-  </data>
-  <data name="UnauthorizedAccess_RegistryNoWrite" xml:space="preserve">
-    <value>Cannot write to the registry key.</value>
-  </data>
   <data name="PersistedFiles_NoHomeDirectory" xml:space="preserve">
     <value>The home directory of the current user could not be determined.</value>
   </data>
